+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 -   Allow construction of Message without body
+-   Ensure Event Handlers are not invoked if there are no PE matches
+    https://github.com/atomist/rug/issues/454
 
 ## [0.17.1] - 2017-03-20